move paconv sa modules to a new file
