+++ mmdet3d/ops/pointnet_modules/__init__.py
@@ -1,8 +1,8 @@
 from .builder import build_sa_module
+from .paconv_sa_module import (PAConvSAModule, PAConvSAModuleCUDA,
+                               PAConvSAModuleMSG, PAConvSAModuleMSGCUDA)
 from .point_fp_module import PointFPModule
-from .point_sa_module import (PAConvSAModule, PAConvSAModuleCUDA,
-                              PAConvSAModuleMSG, PAConvSAModuleMSGCUDA,
-                              PointSAModule, PointSAModuleMSG)
+from .point_sa_module import PointSAModule, PointSAModuleMSG
 
 __all__ = [
     'build_sa_module', 'PointSAModuleMSG', 'PointSAModule', 'PointFPModule',
